Adding the ethers script for the interface contract #45
