--- packages/react-app/src/contracts/ImageRegistry.sol/ImageRegister.dbg.json
@@ -1,4 +1,0 @@
-{
-  "_format": "hh-sol-dbg-1",
-  "buildInfo": "..\\..\\build-info\\0e38ff6f6d19d669a9d0608d51fd5799.json"
-}
