+++ packages/react-app/src/contracts/SBTFactory.sol/SBTFactory.dbg.json
@@ -1,4 +1,4 @@
 {
   "_format": "hh-sol-dbg-1",
-  "buildInfo": "..\\..\\build-info\\06af1486534c120cb0ed8af045422d54.json"
+  "buildInfo": "../../build-info/7bcd0d080771cf9fb269756619fb1690.json"
 }
