--- packages/react-app/src/contracts/Test.sol/Test.dbg.json
@@ -1,4 +1,0 @@
-{
-  "_format": "hh-sol-dbg-1",
-  "buildInfo": "..\\..\\build-info\\63810b9f34ba7bc72c7f81f9c061949f.json"
-}
